update frontend file

+++ RealEstatePrediction_backend/services/Offre/src/main/java/com/example/realestate/prediction/PredictionResponse.java
@@ -5,7 +5,7 @@ package com.example.realestate.prediction;
 public record PredictionResponse(
         String city,
         String date_to_predict,
-        int predicted_price
+        double predicted_price
 ) {
 
 }
